elevate-backend/requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
cryptography==41.0.7

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.0

# Code Execution
requests==2.31.0
httpx==0.25.2

# AI Integration
openai==1.6.1
anthropic==0.8.1

# File Processing
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2

# Monitoring & Logging
sentry-sdk==1.39.2
django-prometheus==2.3.1
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1
factory-boy==3.3.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
faker==20.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

